requirements/base.txt
Django==4.2
django-jazzmin==2.6.0
drf-yasg==1.21.5
djangorestframework==3.14.0
django-cors-headers==3.14.0
django-phonenumber-field==7.1.0
phonenumbers==8.13.9
Pillow==9.4.0
python-environ==0.4.54
environs==9.5.0
psycopg2-binary==2.9.6
pre-commit==3.2.2
django-recaptcha==3.0.0
